DRYing up the startDownload logic. switching over some calls to ApiListener even though they're getting null JSON, a questionable improvement
